+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added SquareRoot method to Calculator class
+- Added Logarithm method to Calculator class
 
 ## [1.0.0] - 2025-09-08